pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobly"
version = "0.1.0"
description = "Automated internship application workflow from SWEList email digests"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "playwright>=1.44.0",
    "sqlmodel>=0.0.18",
    "sqlalchemy>=2.0.0",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.128.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "anthropic>=0.28.0",
    "structlog>=24.1.0",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
jobly = "app.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
